functions/requirements.txt
firebase_functions~=0.1.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
firebase-admin==6.0.1
python-dotenv==1.0.0
openai==1.44.1
